+++ tests/testthat/test-ard_emmeans_mean_difference.R
@@ -17,11 +17,11 @@ test_that("ard_emmeans_mean_difference() works", {
   expect_snapshot(ard_emmeans_mean_difference |> print(columns = "all"))
 
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference, stat_name %in% "method"),
+    cards::get_ard_statistics(ard_emmeans_mean_difference, stat_name %in% "method")[1],
     list(method = "Least-squares adjusted mean difference")
   )
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference, stat_name %in% "estimate") |>
+    cards::get_ard_statistics(ard_emmeans_mean_difference, stat_name %in% "mean.difference.estimate") |>
       unlist() |>
       unname(),
     glm(vs ~ am + mpg, data = mtcars, family = binomial) |>
@@ -29,6 +29,16 @@ test_that("ard_emmeans_mean_difference() works", {
       emmeans::contrast(method = "pairwise") |>
       summary(infer = TRUE) |>
       getElement("estimate")
+  )
+
+  expect_equal(
+    cards::get_ard_statistics(ard_emmeans_mean_difference, stat_name %in% "mean.estimate") |>
+      unlist() |>
+      unname(),
+    glm(vs ~ am + mpg, data = mtcars, family = binomial) |>
+      emmeans::emmeans(specs = ~am, regrid = "response") |>
+      summary(infer = TRUE) |>
+      getElement("prob")
   )
 
 
@@ -45,11 +55,11 @@ test_that("ard_emmeans_mean_difference() works", {
     NA
   )
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference_lme4, stat_name %in% "method"),
+    cards::get_ard_statistics(ard_emmeans_mean_difference_lme4, stat_name %in% "method")[1],
     list(method = "Least-squares mean difference")
   )
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference_lme4, stat_name %in% "estimate") |>
+    cards::get_ard_statistics(ard_emmeans_mean_difference_lme4, stat_name %in% "mean.difference.estimate") |>
       unlist() |>
       unname(),
     lme4::glmer(vs ~ am + (1 | cyl), data = mtcars, family = binomial) |>
@@ -59,6 +69,15 @@ test_that("ard_emmeans_mean_difference() works", {
       getElement("estimate")
   )
 
+  expect_equal(
+    cards::get_ard_statistics(ard_emmeans_mean_difference_lme4, stat_name %in% "mean.estimate") |>
+      unlist() |>
+      unname(),
+    lme4::glmer(vs ~ am + (1 | cyl), data = mtcars, family = binomial) |>
+      emmeans::emmeans(specs = ~am, regrid = "response") |>
+      summary(infer = TRUE) |>
+      getElement("prob")
+  )
 
   #styler: off
   expect_error({
@@ -74,11 +93,11 @@ test_that("ard_emmeans_mean_difference() works", {
   )
   # styler: on
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference_svy, stat_name %in% "method"),
+    cards::get_ard_statistics(ard_emmeans_mean_difference_svy, stat_name %in% "method")[1],
     list(method = "Least-squares mean difference")
   )
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference_svy, stat_name %in% "estimate") |>
+    cards::get_ard_statistics(ard_emmeans_mean_difference_svy, stat_name %in% "mean.difference.estimate") |>
       unlist() |>
       unname(),
     survey::svyglm(api00 ~ sch.wide, design = survey::svydesign(id = ~dnum, weights = ~pw, data = apiclus1, fpc = ~fpc)) |>
@@ -86,6 +105,15 @@ test_that("ard_emmeans_mean_difference() works", {
       emmeans::contrast(method = "pairwise") |>
       summary(infer = TRUE) |>
       getElement("estimate")
+  )
+  expect_equal(
+    cards::get_ard_statistics(ard_emmeans_mean_difference_svy, stat_name %in% "mean.estimate") |>
+      unlist() |>
+      unname(),
+    survey::svyglm(api00 ~ sch.wide, design = survey::svydesign(id = ~dnum, weights = ~pw, data = apiclus1, fpc = ~fpc)) |>
+      emmeans::emmeans(specs = ~sch.wide, regrid = "response") |>
+      summary(infer = TRUE) |>
+      getElement("emmean")
   )
 })