+++ NEWS.md
@@ -1,5 +1,9 @@
 # cardx 0.3.0.9000
 
+* The `ard_complex()` function has been renamed to `ard_mvsummary()`.
+
+* Added mean estimate statistics to the `ard_emmeans_mean_difference()` function.
+
 # cardx 0.3.0
 
 ## New Features and Updates